\name{Endnote2XLS-package}
\alias{Endnote2XLS-package}
\alias{Endnote2XLS}
\docType{package}
\title{
Convert files from Chinese reference database to txt.
}
\description{
Oh, this is a tiny tool just used to convert file format. It is so easy. Welcome to enlarge it.
}
\details{
\tabular{ll}{
Package: \tab Endnote2XLS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-09\cr
License: \tab GPL-2 | GPL-3\cr
}
We delevop such function to convert files extracted from Chinese reference database, such as CNKI, CBM, VIP and Wangfang to text files. Because CBM fail to provide endnote format file, so we develop an option that specify your file whether come from CBM.
}
\author{
KeHao Wu
KeHao Wu<Kehao.wu@gmail.com>
}
